+++ lib/src/clojure_lsp/feature/completion.clj
@@ -440,7 +440,8 @@
                                    (name cursor-value)
                                    (str cursor-value)))
             cursor-full-ns? (when cursor-value-or-ns
-                              (contains? (q/find-all-ns-definition-names analysis) (symbol cursor-value-or-ns)))
+                              (contains? (q/find-all-ns-definition-names analysis)
+                                         (symbol cursor-value-or-ns)))
             items (cond
                     inside-refer?
                     (with-refer-elements matches-fn cursor-loc all-other-ns-elements)