+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Editor
   - Add support to rename namespace of namespaced keywords like re-frame events/subs. #978
   - Improve performance of find-declaration feature. #1021
+  - Fix to avoid suggesting an alias from a clj file to a cljs file. #1024
 
 ## 2022.05.31-17.35.50